==> results/2024_06_16_spectral_preprocessing/02_rce3_calculate_lfp_spectral_alone_comp.ipynb <==
# All oscillation analysis

Brief 1-2 sentence description of notebook.

import warnings
warnings.filterwarnings('ignore')

import os
import collections
import itertools
from collections import defaultdict
from itertools import combinations

os.environ["SPECTRAL_CONNECTIVITY_ENABLE_GPU"] = "true"
import cupy as cp

# Imports of all used packages and libraries
import numpy as np
import pandas as pd
from scipy import stats


import matplotlib
import matplotlib.pyplot as plt

from spectral_connectivity import Multitaper, Connectivity
import spectral_connectivity

from scipy.signal import butter, filtfilt, hilbert

# Functions

def generate_pairs(lst):
    """
    Generates all unique pairs from a list.

    Parameters:
    - lst (list): The list to generate pairs from.

    Returns:
    - list: A list of tuples, each containing a unique pair from the input list.
    """
    n = len(lst)
    return [(lst[i], lst[j]) for i in range(n) for j in range(i+1, n)]

def update_array_by_mask(array, mask, value=np.nan):
    """
    Update elements of an array specified by a mask with a given value.
    
    Parameters:
        array (np.ndarray): The original array to be modified.
        mask (np.ndarray): A boolean array of the same shape as `array`.
                           Elements of `array` corresponding to True in the `mask`
                           will be updated.
        value (scalar, optional): The value to assign to elements of `array` specified by `mask`.
                                  Default is np.nan.
    
    Returns:
        np.ndarray: A copy of `array` with the specified elements updated.
    """
    result = array.copy()  # Create a copy of the input array to modify
    result[mask] = value   # Apply the mask to update the designated elements
    return result

## Inputs & Data

Explanation of each input and where it comes from.

# Inputs and Required data loading
# input varaible names are in all caps snake case
# Whenever an input changes or is used for processing 
# the vairables are all lower in snake case
OUTPUT_DIR = r"./proc" # where data is saved should always be shown in the inputs

TIME_HALFBANDWIDTH_PRODUCT = 2
TIME_WINDOW_DURATION = 1
TIME_WINDOW_STEP = 0.5
RESAMPLE_RATE=1000

zscore_threshold = 4
VOLTAGE_SCALING_VALUE = 0.195

FILTER_BAND = (0, 56)

OUTPUT_PREFIX = "rce_pilot_3_alone_comp"

LFP_TRACES_DF = pd.read_pickle("./proc/{}_01_lfp_traces_and_frames.pkl".format(OUTPUT_PREFIX))

LFP_TRACES_DF.shape

## Preprocessing

original_trace_columns = [col for col in LFP_TRACES_DF.columns if "trace" in col]

original_trace_columns

for col in original_trace_columns:
    print(col)
    LFP_TRACES_DF[col] = LFP_TRACES_DF[col].apply(lambda x: x.astype(np.float32) * VOLTAGE_SCALING_VALUE)

LFP_TRACES_DF.head()

# Calculating modified zscore

`0.6745(xi – x̃) / MAD`

for col in original_trace_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_MAD".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: stats.median_abs_deviation(x))

for col in original_trace_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_modified_zscore".format(brain_region)
    MAD_column = "{}_lfp_MAD".format(brain_region)

    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF.apply(lambda x: 0.6745 * (x[col] - np.median(x[col])) / x[MAD_column], axis=1)

LFP_TRACES_DF[updated_column]

## calculating root mean sequare

for col in original_trace_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_RMS".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: (x / np.sqrt(np.mean(x**2))).astype(np.float32))


LFP_TRACES_DF.head()

## Filtering for zscore value

zscore_columns = [col for col in LFP_TRACES_DF.columns if "zscore" in col]

zscore_columns

for col in zscore_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_mask".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: np.abs(x) >= zscore_threshold)

LFP_TRACES_DF[updated_column].head()

LFP_TRACES_DF[updated_column].iloc[0].shape

sum(LFP_TRACES_DF[updated_column].iloc[0])

- Filtering raw traces by zscore

LFP_TRACES_DF[col].head()

for col in original_trace_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_trace_filtered".format(brain_region)    
    mask_column = "{}_lfp_mask".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF.apply(lambda x: update_array_by_mask(x[col], x[mask_column]), axis=1)

LFP_TRACES_DF[col].head()

sum(np.isnan(LFP_TRACES_DF[col].iloc[0]))

- Calculating RMS of filtered signal

filtered_trace_column = [col for col in LFP_TRACES_DF if "lfp_trace_filtered" in col]

for col in filtered_trace_column:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_lfp_RMS_filtered".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: (x / np.sqrt(np.nanmean(x**2))).astype(np.float32))

plt.plot(LFP_TRACES_DF[original_trace_columns[0]].iloc[0])

plt.plot(LFP_TRACES_DF["{}_lfp_RMS".format(brain_region)].iloc[0])

plt.plot(LFP_TRACES_DF["{}_lfp_RMS_filtered".format(brain_region)].iloc[0])

## Calculating phase of signals

- Filtering for theta and gamma

RMS_columns = [col for col in LFP_TRACES_DF if "RMS" in col and "filtered" not in col]

fs = 1000
order=4

freq_band = [4, 12]
b, a = butter(order, freq_band, fs=fs, btype='band')

for col in RMS_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_theta_band".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: filtfilt(b, a, x, padtype=None).astype(np.float32))

freq_band = [30, 50]
b, a = butter(order, freq_band, fs=fs, btype='band')

for col in RMS_columns:
    print(col)
    brain_region = col.split("_")[0]
    updated_column = "{}_gamma_band".format(brain_region)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: filtfilt(b, a, x, padtype=None).astype(np.float32))

- Calculating the phase

band_columns = [col for col in LFP_TRACES_DF if "band" in col]

band_columns

for col in band_columns:
    print(col)
    brain_region = col.replace("_band", "")
    updated_column = "{}_phase".format(brain_region)
    print(updated_column)
    LFP_TRACES_DF[updated_column] = LFP_TRACES_DF[col].apply(lambda x: np.angle(hilbert(x), deg=False))

LFP_TRACES_DF[col]

# Dropping unused columns

list(LFP_TRACES_DF.columns)

dropping_columns = [col for col in LFP_TRACES_DF if ("MAD" in col or "modified_zscore" in col or "RMS" in col or "mask" in col or "trace_filtered" in col) and "RMS_filtered" not in col]




dropping_columns

LFP_TRACES_DF = LFP_TRACES_DF.drop(columns=dropping_columns, errors="ignore")

# Power Calcuation

- Getting the column name of all the traces

input_columns = [col for col in LFP_TRACES_DF.columns if "RMS" in col and "filtered" in col]

input_columns

for col in input_columns:
    print(col)
    LFP_TRACES_DF[col] = LFP_TRACES_DF[col].apply(lambda x: np.round(x.astype(np.float32), decimals=3))

- Calcuating the power at each frequency band

LFP_TRACES_DF[col].iloc[0]

for col in input_columns:
    # brain_region = col.split("_")[0]
    brain_region = col.replace("_lfp", "").replace("RMS_filtered", "").strip("_")
    print(brain_region)

    try:
        # Define column names
        frequencies_col = f"{brain_region}_power_frequencies"
        power_col = f"{brain_region}_power_all_frequencies_all_windows"

        index_to_multitaper = {}
        index_to_connectivity = {}
        index_to_frequencies = {}
        index_to_power = {}

        # Apply Multitaper function to the lfp_trace column
        for index, row in LFP_TRACES_DF.iterrows():
            print(index)
            index_to_multitaper[index] = Multitaper(
                    time_series=row[col], 
                    sampling_frequency=RESAMPLE_RATE, 
                    time_halfbandwidth_product=TIME_HALFBANDWIDTH_PRODUCT,
                    time_window_duration=TIME_WINDOW_DURATION, 
                    time_window_step=TIME_WINDOW_STEP
                )
            index_to_connectivity[index] = Connectivity.from_multitaper(index_to_multitaper[index])
            index_to_frequencies[index] = index_to_connectivity[index].frequencies[FILTER_BAND[0]: FILTER_BAND[1]]
            index_to_power[index] = index_to_connectivity[index].power().squeeze()[:, FILTER_BAND[0]: FILTER_BAND[1]]

        LFP_TRACES_DF[frequencies_col] = LFP_TRACES_DF.index.to_series().map(index_to_frequencies)
        LFP_TRACES_DF[power_col] = LFP_TRACES_DF.index.to_series().map(index_to_power)
        LFP_TRACES_DF[power_col] = LFP_TRACES_DF[power_col].apply(lambda x: x.astype(np.float32))
    
    except Exception as e: 
        print(e)
        

- Getting the timestamps of the power

LFP_TRACES_DF["power_timestamps"] = LFP_TRACES_DF["lfp_timestamps"].apply(lambda x: x[(RESAMPLE_RATE//2):(-RESAMPLE_RATE//2):(RESAMPLE_RATE//2)])
# .iloc[0][500:-500:500].shape

- Making sure that the timestamps for power makes sense with shape and values

LFP_TRACES_DF["power_timestamps"].head().apply(lambda x: x.shape)

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "power_all_frequencies_all_windows" in col][0]].iloc[0].shape

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "lfp_timestamps" in col][0]].iloc[0]

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "power_timestamps" in col][0]].iloc[0]

- Checking if the right frequencies are being used

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "power_frequencies" in col]].head()

LFP_TRACES_DF["power_calculation_frequencies"] = LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "power_frequencies" in col][0]].copy()

- Dropping unnecessary columns

LFP_TRACES_DF = LFP_TRACES_DF.drop(columns=[col for col in LFP_TRACES_DF.columns if "power_frequencies" in col], errors="ignore")

LFP_TRACES_DF.head()

## Coherence Calculation

- Getting the trace column pairs

brain_region_pairs = generate_pairs(sorted(input_columns))
brain_region_pairs = sorted(brain_region_pairs)


brain_region_pairs

## Coherece Calculation

- Calculating the coherence

for region_1, region_2 in brain_region_pairs:
    # Define base name for pair
    pair_base_name = f"{region_1.split('_')[0]}_{region_2.split('_')[0]}"
    print(pair_base_name)

    try:
        # Define column names
        frequencies_col = f"{pair_base_name}_coherence_frequencies"
        coherence_col = f"{pair_base_name}_coherence_all_frequencies_all_windows"

        index_to_multitaper = {}
        index_to_connectivity = {}
        index_to_frequencies = {}
        index_to_coherence = {}

        for index, row in LFP_TRACES_DF.iterrows():
            index_to_multitaper[index] = Multitaper(
                time_series=np.array([row[region_1], row[region_2]]).T, 
                sampling_frequency=RESAMPLE_RATE, 
                time_halfbandwidth_product=TIME_HALFBANDWIDTH_PRODUCT, 
                time_window_step=TIME_WINDOW_STEP, 
                time_window_duration=TIME_WINDOW_DURATION
            )
            print(index)
            index_to_connectivity[index] = Connectivity.from_multitaper(index_to_multitaper[index])
            index_to_frequencies[index] = index_to_connectivity[index].frequencies[FILTER_BAND[0]: FILTER_BAND[1]]
            index_to_coherence[index] = index_to_connectivity[index].coherence_magnitude()[:, FILTER_BAND[0]: FILTER_BAND[1], 0, 1]

        LFP_TRACES_DF[frequencies_col] = LFP_TRACES_DF.index.to_series().map(index_to_frequencies)
        LFP_TRACES_DF[coherence_col] = LFP_TRACES_DF.index.to_series().map(index_to_coherence)
        LFP_TRACES_DF[coherence_col] = LFP_TRACES_DF[coherence_col].apply(lambda x: x.astype(np.float32))

    except Exception as e: 
        print(e)

LFP_TRACES_DF["BLA_LH_coherence_all_frequencies_all_windows"].iloc[0].shape

LFP_TRACES_DF["BLA_LH_coherence_frequencies"].iloc[0]

LFP_TRACES_DF.head()

- Getting the timestamps of the coherence

LFP_TRACES_DF["coherence_timestamps"] = LFP_TRACES_DF["lfp_timestamps"].apply(lambda x: x[(RESAMPLE_RATE//2):(-RESAMPLE_RATE//2):(RESAMPLE_RATE//2)])


- Making sure that the timestamps for coherence makes sense with shape and values

LFP_TRACES_DF["coherence_timestamps"].head().apply(lambda x: x.shape)

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "coherence_all_frequencies_all_windows" in col][0]].iloc[0].shape

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "lfp_timestamps" in col][0]].iloc[0]

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "coherence_timestamps" in col][0]].iloc[0]

- Checking if the right frequencies are being used

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "coherence_frequencies" in col]].head()

LFP_TRACES_DF["coherence_calculation_frequencies"] = LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "coherence_frequencies" in col][0]].copy()

- Dropping unnecessary columns

LFP_TRACES_DF = LFP_TRACES_DF.drop(columns=[col for col in LFP_TRACES_DF.columns if "coherence_frequencies" in col], errors="ignore")

LFP_TRACES_DF.head()

# Calculate Granger's

for region_1, region_2 in brain_region_pairs:
    # Define base name for pair
    region_1_base_name = region_1.split('_')[0]
    region_2_base_name = region_2.split('_')[0]

    pair_base_name = f"{region_1_base_name}_{region_2_base_name}"
    print(pair_base_name)

    try:
        # Define column names
        frequencies_col = f"{pair_base_name}_granger_frequencies"
        granger_1_2_col = f"{region_1_base_name}_{region_2_base_name}_granger_all_frequencies_all_windows"
        granger_2_1_col = f"{region_2_base_name}_{region_1_base_name}_granger_all_frequencies_all_windows"

        index_to_multitaper = {}
        index_to_connectivity = {}
        index_to_frequencies = {}
        index_to_granger_1_to_2 = {}
        index_to_granger_2_to_1 = {}

        
        for index, row in LFP_TRACES_DF.iterrows():
            # Apply Multitaper function
            index_to_multitaper[index] = Multitaper(
                    time_series=np.array([row[region_1], row[region_2]]).T, 
                    sampling_frequency=RESAMPLE_RATE, 
                    time_halfbandwidth_product=TIME_HALFBANDWIDTH_PRODUCT, 
                    time_window_step=TIME_WINDOW_STEP, 
                    time_window_duration=TIME_WINDOW_DURATION
                )
            
            print(index)
            index_to_connectivity[index] = Connectivity.from_multitaper(index_to_multitaper[index])
            index_to_frequencies[index] = index_to_connectivity[index].frequencies[FILTER_BAND[0]: FILTER_BAND[1]]
            index_to_granger_1_to_2[index] = index_to_connectivity[index].pairwise_spectral_granger_prediction()[:, FILTER_BAND[0]: FILTER_BAND[1], 0, 1]
            index_to_granger_2_to_1[index] = index_to_connectivity[index].pairwise_spectral_granger_prediction()[:, FILTER_BAND[0]: FILTER_BAND[1], 1, 0]          
            

        LFP_TRACES_DF[frequencies_col] = LFP_TRACES_DF.index.to_series().map(index_to_frequencies)
        LFP_TRACES_DF[granger_1_2_col] = LFP_TRACES_DF.index.to_series().map(index_to_granger_1_to_2)
        LFP_TRACES_DF[granger_1_2_col] = LFP_TRACES_DF[granger_1_2_col].apply(lambda x: x.astype(np.float32))
        LFP_TRACES_DF[granger_2_1_col] = LFP_TRACES_DF.index.to_series().map(index_to_granger_2_to_1)
        LFP_TRACES_DF[granger_2_1_col] = LFP_TRACES_DF[granger_2_1_col].apply(lambda x: x.astype(np.float32))        
    
    except Exception as e: 
        print(e)

- Getting the timestamps of the granger

LFP_TRACES_DF["granger_timestamps"] = LFP_TRACES_DF["lfp_timestamps"].apply(lambda x: x[(RESAMPLE_RATE//2):(-RESAMPLE_RATE//2):(RESAMPLE_RATE//2)])


- Making sure that the timestamps for granger makes sense with shape and values

LFP_TRACES_DF["granger_timestamps"].head().apply(lambda x: x.shape)

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "granger_all_frequencies_all_windows" in col][0]].iloc[0].shape

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "lfp_timestamps" in col][0]].iloc[0]

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "granger_timestamps" in col][0]].iloc[0]

- Checking if the right frequencies are being used

LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "granger_frequencies" in col]].head()

LFP_TRACES_DF["granger_calculation_frequencies"] = LFP_TRACES_DF[[col for col in LFP_TRACES_DF.columns if "granger_frequencies" in col][0]].copy()

LFP_TRACES_DF = LFP_TRACES_DF.drop(columns=[col for col in LFP_TRACES_DF.columns if "granger_frequencies" in col], errors="ignore")

LFP_TRACES_DF.head()

# Saving to file

LFP_TRACES_DF.head()

LFP_TRACES_DF.columns

LFP_TRACES_DF.to_pickle("./proc/{}_02_full_spectral.pkl".format(OUTPUT_PREFIX))

raise ValueError()
